Set proper extension on file attachment download
On attachment download, file extensions were no longer correctly set.
This was inconvenient because file extension could not be properly set.
For instance, instead of downloading "myattachment.pdf", the filename
was set to "hvriuehvrhjzkjpeofu1345RFV76cgfc" (without the ".pdf"
suffix).

This regression was introduced when file keys stopped referring to file
names. Now, we can no longer retrieve the original file name, just the
extension.

Note that the proposed fix works only for the "filesystem" backend: for
the "django" backend, we do not store the metadata associated to the
uploaded files. Thus, the original file content type is lost after the
upload.

This issue was raised a couple times publicly, although no formal issue
was created in the github repo:
2017-03: https://groups.google.com/forum/#!topic/edx-code/YJuNMeIaJi0
2018-03: https://openedx.slack.com/messages/C08B4LZEZ/convo/C08B4LZEZ-1521550202.000508/?thread_ts=1521550202.000508
2018-09: https://openedx.slack.com/messages/C08B4LZEZ/convo/C08B4LZEZ-1537270442.000100/